update: forbidden error about updating and deleting post when user uuid is not matched

+++ src/post/post.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { PostRepository } from './post.repository';
 import { PostDto } from './dto/post.dto';
 import { CreatePostDto } from './dto/createPost.dto';
@@ -6,6 +6,7 @@ import { UpdatePostDto } from './dto/updatePost.dto';
 import { PostFullContent } from './types/postFullContent';
 import { PostListQueryDto } from './dto/postListQuery.dto';
 import { PostListDto } from './dto/postList.dto';
+import { PayloadDto } from 'src/auth/dto/payload.dto';
 
 @Injectable()
 export class PostService {
@@ -51,12 +52,24 @@ export class PostService {
     return this.getPost(post.uuid);
   }
 
-  async updatePost(uuid: string, postDto: UpdatePostDto): Promise<PostDto> {
+  async updatePost(
+    uuid: string,
+    postDto: UpdatePostDto,
+    user: PayloadDto,
+  ): Promise<PostDto> {
+    const authorId = (await this.postRepository.getPost(uuid)).authorId;
+    if (authorId !== user.uuid)
+      throw new ForbiddenException('Not match user uuid');
+
     await this.postRepository.updatePost(uuid, postDto);
     return this.getPost(uuid);
   }
 
-  async deletePost(uuid: string): Promise<PostFullContent> {
+  async deletePost(uuid: string, user: PayloadDto): Promise<PostFullContent> {
+    const authorId = (await this.postRepository.getPost(uuid)).authorId;
+    if (authorId !== user.uuid)
+      throw new ForbiddenException('Not match user uuid');
+
     return await this.postRepository.deletePost(uuid);
   }
 }
